requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
opencv-python
numpy
torch
torchvision
faiss-cpu
pillow
transformers
opencv-python-headless
google-cloud-storage
python-dotenv
google-generativeai
huggingface_hub
timm
scikit-learn
requests
redis
passlib[bcrypt]
motor
pymongo
python-jose[cryptography]
aiosmtplib
pydantic[email]
google-cloud-secret-manager
python-multipart
bcrypt
huggingface_hub